requirements.txt
asgiref==3.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
Django==3.2.16
flake8==5.0.4
idna==3.11
mccabe==0.7.0
pillow==11.3.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==1.2.1
pytz==2022.6
requests==2.32.5
sqlparse==0.4.3
typing_extensions==4.15.0
urllib3==2.5.0
